Bunch of refactoring. Other fixes. Yeah, really specific, I know.
Former-commit-id: 8977553819f39c9c74312eca2771606ae763afb5

+++ app/models/database/queries/ddl/CreateDailyMetricsTable.scala
@@ -8,7 +8,7 @@ case object CreateDailyMetricsTable extends Statement {
     (
        day date not null,
        metric character varying(128) not null,
-       value integer not null default 0,
+       value bigint not null default 0,
        measured timestamp without time zone not null,
        constraint pk_daily_metrics primary key (day, metric)
     ) with (oids = false);
